+++ lib/models/quest_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:g_sui_hunter/quest.dart';
+import 'package:g_sui_hunter/models/quest.dart';
 
 class QuestModel extends ChangeNotifier {
   List<Quest> questList = [];
